exporter/api/conn: drop test-only nil-ws guard from Close
Close had a nil check only because TestConn_Send_FullQueue passed
nil for ws. Use a real *websocket.Conn (dialed against an httptest
echo) so the guard can go and Close stays a single straight line.

Also move TestConn_Send_FullQueue and the dialTestWebsocket helper
out of broadcaster_test.go into a new conn_test.go — they're conn
contract tests, not broadcaster ones.
